a whole bunch of changes, remove term from fibonacci, change solve order

+++ src/main.rs
@@ -4,7 +4,7 @@ mod sequence;
 use fraction::GenericFraction;
 use sequence::Sequence;
 use std::str::FromStr;
-pub type Term = GenericFraction<i64>;
+pub type Term = GenericFraction<i128>;
 
 use clap::Parser;
